Cygwin: Force keyboard state onto VCK at startup
Copy the state of the Windows keyboard device to the Virtual Core Keyboard at startup.

Otherwise, this happens lazily after the first keypress, which can lead to applications
which are started from a shell window and inspect the keyboard state before a character is
typed getting the wrong idea about the desired keymap (e.g. xemacs shows this behaviour)

Signed-off-by: Jon TURNEY <jon.turney@dronecode.org.uk>