Hint handling fixes
Respect the systems ownership of the clipping region

When the style changes, adjust the window size so the client area remains the same.
Otherwise the window size may change when sizing is reflected from
Windows to X, and some windows are drawn expecting them to be exactly the reqeusted size
(e.g. gmplayer control window)

Use the current style in determining the window border when validating which sizing
XXX: Needs to allow for different border thickness when uncaptioned?

Signed-off-by: Jon TURNEY <jon.turney@dronecode.org.uk>