Replace DDXBEFORERESET with a more general way of doing DDX-specific hooks
Replace DDXBEFORERESET, which was introduced for the sole benefit of XWin
in commit 048045a9, with a slightly more general way of doing DDX-specific
hooks, still for the sole benefit of XWin :-)

Use a structure of function pointers, which can be initialized as required
by the specific DDX during main(), before dix_main() is called.

Change main() in the XWin DDX to initializes the ddxBeforeReset hook.

Signed-off-by: Jon TURNEY <jon.turney@dronecode.org.uk>