Replace DDXBEFORERESET with a more general way of doing DDX-specific hooks
Replace DDXBEFORERESET, which was introduced for the sole benefit of XWin
in commit 048045a9, with a slightly more general way of doing DDX-specific
hooks, still for the sole benefit of XWin :-)

Use a structure of function pointers, which can be initialized as required
by the specific DDX

Add DDXMAIN to control if the DIX calls ddxMain(), which can be used by
the DDX to initialize ddxHooks as required.

Provide a ddxMain() in the XWin DDX which initializes the ddxBeforeReset
hook.

Unfortunately, this retains the behaviour that all DDX built at the same
time must provide that function, whether they need it or not, but there
doesn't seem to be a clean way around that.

Signed-off-by: Jon TURNEY <jon.turney@dronecode.org.uk>