Add a DDX specific GLX provider push hook
This is needed to give DDX which are statically linked with extensions
a chance to install DDX specific GLX providers before the GLX extension
is initialized

The the swrast provider is installed just before GLX extension is
initialized.  The GLX extension asks providers if they can support a
screen in the reverse of the provider installation order, so installing
a DDX-specific GLX provider earlier than this is not useful, as it will
always lose out to swrast.  Installing the provider later than this is
not useful as the provider for each screen has already been selected
during GLX extension initialization.

Signed-off-by: Jon TURNEY <jon.turney@dronecode.org.uk>