Do not create visuals from fbConfigs which don't support window drawables.
It is implicit that visuals are suitable for use with a window.

Glean (for e.g.) assumes all visuals support window drawables.

glXChooseVisual() documentation also says:
"There is no direct filter for picking only visuals that support GLXPixmaps.
GLXPixmaps are supported for visuals whose GLX_BUFFER_SIZE  is one of the
pixmap depths supported by the X server."

This appears to perhaps mean that we should also only create visuals from
fbConfigs which have a GLX_BUFFER_SIZE which is one of the Pixmap depths
supported by the X server if drawableType contains GLX_PIXMAP_BIT.

Signed-off-by: Jon TURNEY <jon.turney@dronecode.org.uk>