Cygwin/X: Avoid cursor size log spam
Fedora 12 likes to use a 39x26 animated wait cursor.  Avoid spamming
the log with warnings that each frame can't be completely contained
in the 32x32 native cursor

Also reformat log message so it doesn't contain a '\n\t'. I mean,
it's not like we might want to grep the log or something...

Signed-off-by: Jon TURNEY <jon.turney@dronecode.org.uk>