Cygwin/X: Don't terminate Xserver if clipboard client is killed
This is a temporary solution.

Really we should either improve the way we try to avoid the internal
clipboard client getting killed or perhaps restart it after an interval.

Avoiding getting killed by the XDM is horribly heuristic at the moment,
as we basically try to guess about what the XDM is doing, so this is
always going to be fragile. (Should just check that we can't actually
intercept the close request for the client)

Restarting the clipboard client thread is complicated by the fact that
at the moment the clipboard thread exits on server regeneration and is
restarted by hooks installed by InitInput())
(So we either need to make arrangements not restart when server is restarting,
or only start once and properly re-cache atoms on server regeneration)

Also: http://xkcd.com/292/

Signed-off-by: Jon TURNEY <jon.turney@dronecode.org.uk>