Define FD_SETSIZE on Cygwin regardless of XWin DDX
All DDXs segfault on Cygwin unless -DFD_SETSIZE=256 is set, so make sure
we do so whether or not we are building XWin.

FD_SETSIZE must be at least XFD_SETSIZE for uses of select() to be correct.
The Cygwin default is only 64, so it must be increased to 256

Signed-off-by: Yaakov Selkowitz <yselkowitz@users.sourceforge.net>
Signed-off-by: Jon TURNEY <jon.turney@dronecode.org.uk>
Reviewed-by: Colin Harrison <colin.harrison@virgin.net>
(cherry picked from commit ed2c2a68e1e6b409c198a52d99d300eb6d517e89)