Remove any old logfile before trying to write to it
If we are not backing up logfiles, remove the old logfile before
trying to write a new logfile, as otherwise the operation may fail
if the previous logfile was created by a different user.

This looks like a generic problem, but probably isn't seen with
the Xfree86 DIX as that uses the logfile backup mechanism.

Signed-off-by: Jon TURNEY <jon.turney@dronecode.org.uk>
Acked-by: Yaakov Selkowitz <yselkowitz@users.sourceforge.net>
Signed-off-by: Yaakov Selkowitz <yselkowitz@users.sourceforge.net>