Cygwin/X: Cause the X server to terminate if clipboard or WM internal client threads exit due to an error
Calling FatalError from an internal client thread is a no-no, so these
failures are silent at the moment (in the sense that no pop-up occurs to
tell the user that we are terminating or why).

Really we should either improve the way we try to avoid the internal
clipboard client getting killed or perhaps restart it after an interval.

Avoiding getting killed by the XDM is horribly heuristic at the moment,
as we basically try to guess about what the XDM is doing, so this is
always going to be fragile. (Should just check that we can't actually
intercept the close request for the client)

Restarting the clipboard client thread is complicated by the fact that
at the moment the clipboard thread exits on server regeneration and is
restarted by hooks installed by InitInput())
(So we either need to make arrangements not restart when server is restarting,
or only start once and properly re-cache atoms on server regeneration)

Also, I'm just noticing I should be using pthread_kill() to ensure the
signal is delivered to the thread which is running the main dispatch loop
so that will exit the select() it may be blocked in and notice that
DE_TERMINATE is set.

Also: http://xkcd.com/292/

Signed-off-by: Jon TURNEY <jon.turney@dronecode.org.uk>
Signed-off-by: Yaakov Selkowitz <yselkowitz@users.sourceforge.net>