Improve choice of display name used by internal clients
Choose display name used to connect internal clients and exported into
environment of processes started by traymenu so that it uses a transport we know
is working

This should mean the server can start correctly with -multiwindow and/or
-clipboard and any two of -nolisten inet6, -nolisten inet and -nolisten unix
(the server will correctly refuse to start if all 3 are used, as it must be
listening on at least one socket)

v2:
Place prototype for winGetDisplayName() in windisplay.h, and include it where
needed.

XXX: TransIsListening would be better written in xtrans, rather than grovelling
around in the servers list of listeners, see [1] for patch.

[1] https://sourceware.org/bugzilla/show_bug.cgi?id=10725

Signed-off-by: Jon TURNEY <jon.turney@dronecode.org.uk>
Reviewed-by: Yaakov Selkowitz <yselkowitz@users.sourceforge.net>