hw/xwin: Make winOverrideIcon() thread-safe for icon data access
winOverrideIcon() is called from the internal WM client thread.

Accessing server-internal data structures to get icon data or
window hints is not safe, as there is no lock to ensure we do not collide
with these data structures being updated in the server thread.

Rewrite so the internal client thread uses X client calls to
obtain this data safely

We used to also set the icon inside the server when the window was
initially created.  For simplicity, we simply send a message to the
internal WM to update the icon when the window is created (rather
than writing different icon update code which can work in the server
thread for that one case...)

extwm mode used to do the icon update in the server. I'm not sure that
actually made much sense.  Let's assume the external WM client can do it
instead...

Signed-off-by: Jon TURNEY <jon.turney@dronecode.org.uk>