Let configure --enable/disable-aiglx control building of AIGLX for all DDXs
Let configure --enable/disable-aiglx control building of AIGLX for all DDXs. Currently
we can't use --enable/disable-aiglx to control if Xwin DDX is built with AIGLX enabled,
as at the moment it's forced off if we aren't building the X.Org DDX DRI loader

Rearrange things a bit, introducing a new configure symbol, AIGLX_DRI_LOADER to
specifically indicate if the X.Org DDX DRI loader should be built, and replace the previous
X.Org DDX-specific uses of AIGLX with that.

As before, AIGLX_DRI_LOADER is only enabled if all of --enable-glx, --enable-aiglx and
--enable-dri are

This allows the general symbol AIGLX to control if AIGLX is built for the XWin DDX as
well

Perhaps some more refactoring could be done here to make the configuration of DRI1 and
DRI2 loaders orthogonal, so they can be configured independently.

AC_DEFINE(AIGLX) seems to be redundant, I can't find anything which checks it

Signed-off-by: Jon TURNEY <jon.turney@dronecode.org.uk>