Cygwin/X: Linkage bodge for swrast, part 1
Remove glapi.c and glthread.c, which contain GL dispatcher symbols
which are defined in libGL, and link with libGL for them instead, as
dri_swrast.so is linked with libGL on Cygwin to avoid undefined symbols
at link time.

Also remove glprocs.h which is only referred to by those .c files.