Cygwin/X: Preserve client area size on Windows window style change
(This replaces the previous "Hint handling fixes" patch

When the style changes, adjust the window size so the client area remains the same.
Otherwise the window size may change when sizing is reflected from
Windows to X, and some windows are drawn expecting them to be exactly the requested size
(e.g. gmplayer control window)

As a consquence of this, we need to be more careful to create windows with exactly
the requested client area initially.

Signed-off-by: Jon TURNEY <jon.turney@dronecode.org.uk>