os: Use external tool for creating backtraces on crashes if available.
Make the code for fork/exec-ing a helper program to perform a backtrace more
generic, so it can be used to call /usr/bin/xorg-backtrace, a script which
invokes gdb to create reasonable backtraces. On errors it falls back to the
generic method.

v2: disable backtrace code on MinGW

Based on a patch by Matthias Hopf mhopf at suse.de

Signed-off-by: Jon TURNEY <jon.turney@dronecode.org.uk>
Signed-off-by: Yaakov Selkowitz <yselkowitz@users.sourceforge.net>