When XDMCP -from is specified, only register the requested address
When XDMCP -from is specified, only register the requested address,
rather than the requested address, and any others we have of different
address families to the requested address.

e.g. if we have 4 interfaces with both IPv4 and IPv6 addresses (which
are not IPv6 mapped IPV4 addresses), using -from with one of those IPv4
addresses means only that IPv4 address, and all IPv6 addresses are used
in the connection data in XDMCP REQUEST packet.

Signed-off-by: Jon TURNEY <jon.turney@dronecode.org.uk>