Consolidate XInitThreads/XSupportsLocale() in a single place.
Consolidate XInitThreads/XSupportsLocale() in a single place.
Also, if XSupportsLocale() fails, set a known good locale.

This needs cleanup. Getting the prototypes for libX11 functions in
scope in server startup code is a problem.

v2:
Remove winclipboardProc_Exit label which is now unused

v3:
When reporting locale is not supported by X, log the actual locale