Allow XScreenSaverSuspend() to effect Windows screen saver or monitor power off
Reset the idle timer in the block handler if screenSaverSuspended
is set.

This isn't quite a complete solution.  We should also set the block
timeout to something less than the idle timer timeout to ensure we
will reset the idle timer before it times out if we are blocking.

The idle timer timeout is presumably the first one to expire of the
screen saver activation or monitor low power or power down timeout,
depending on configuration.

Unfortunately this is probably not straightforward to do. Whilst
SystemParametersInfo(SPI_GETSCREENSAVETIMEOUT) is portable, apparently
SPI_GETLOWPOWERTIMEOUT and SPI_GETPOWEROFFTIMEOUT are not supported
by Windows versions 6.0 or later, and the interface for discovering
equivalent value is complex.

This doesn't matter in the case where a media player or similar
application is the one making the XScreenSaverSuspend() requests, as
it will be continuously drawing, causing the X server to become
unblocked often.

In the case where slide show presentation application or similar
is the one making the XScreenSaverSuspend() request, this might be
a problem.

Since "Powerpoint is Evil" [1],  I think we'll leave it like this
till someone complains :-)

[1] http://www.edwardtufte.com/tufte/books_pp