Improve WM_ENDSESSION handling using separate messaging window thread
WM_ENDSESSION just calls GiveUp to set the DE_TERMINATE flag.  We also need the
X server dispatch loop to be unblocked so it can notice that DE_TERMINATE has
been set and the server exits cleanly, removing it's lock file and any unix
domain socket.

The system will terminate the process when the last UI thread in that process
returns from processing WM_ENDSESSION for the last top-level window.

Since WM_ENDSESSION appears to sent by the system via SendMessage(), which means
the wndproc is called to process it directly from inside a blocking syscall in
the message thread for that window (the X sever thread), we can't easily terminate
the dispatch loop from inside the message processing.

So, create a messaging window, a hidden, top-level window, with a separate
thread to catch this message, and process it, blocking until the X server
dispatch loop exits.

Also, notice when this is a shutdown cancel WM_ENDSESSION message and take
no action.

Signed-off-by: Jon TURNEY <jon.turney@dronecode.org.uk>