Add ' (on <WM_CLIENT_MACHINE>)' to window name when it's useful to do so
Enhance GetWindowName() so it appends the result of XGetWMClientMachine()
when it is avilable and useful to do so

Refactor Xutf8TextPropertyToString() from GetWindowName() as a separate
utility function

Simplify GetWindowName() by moving UTF-8 to whcar conversion out to it's
call site

Add -hostintitle option to control this behaviour.  Add documentation for
this option to man page and -help text.

Signed-off-by: Jon TURNEY <jon.turney@dronecode.org.uk>