hw/xwin/glx: Blacklist 'GDI generic' GL renderer
If the native GL renderer is the GDI generic renderer, (as can happen
if we are in safe mode or the video driver is VGA), don't use it.
It's not accelerated and we will probably get better conformance and
perfomance from swrast.

Don't install screen function wrappers in glxWinScreenProbe unless
we are succesful.

Also, move fbConfig dumping to after GLX version has been determined
from extensions

Signed-off-by: Jon TURNEY <jon.turney@dronecode.org.uk>