Empty Windows clipboard when X loses focus
At the moment, we only notice selection updates when the selection owner
changes. So, unfortunately, applications which already own a selection can
change that selection without us noticing.

For example, gitk (and probably all applications using Tk) only claim
the selection when it moves between controls, one can adjust the selection
within a control without any notice.

This simple fix assumes that whenever X gains focus, the selection will be
adjusted, so when we lose focus, we should clear the Windows clipboard.

This ensures Windows will ask for the contents to be rendered again if Windows
application asks for the content, ensuring they get the updated content.

(This is almost, but not quite right.  If the selection isn't altered
while we have the focus, then we've destroyed what was on the clipboard
previously.)

Signed-off-by: Jon TURNEY <jon.turney@dronecode.org.uk>