Introduce winProcessXEventsTimeout() to the concept of fractions of a second
Oh this is terrible.

Currently we only compute the select timeout in whole seconds.  This means if
we have less than 1 second remaining, we select with a timeout of 0 (i.e. poll)
which causes the task to spin, burning 100% CPU for the remaining timeout (and
possibly preventing the process we are waiting for from running :S)