Use spawnl() rather than pipe() & fork() to invoke xkbcomp
Implement System() specially on Cygwin not to use fork() and exec(), but
spawnl()

Use System() and a temp file rather than pipe() & fork() to invoke xkbcomp on
cygwin, just as done on mingw

Somewhat the counsel of despair, but sometimes Windows is being such a little
bitch that the cygwin DLL just can't emulate fork() successfully, so just using
spawnl() here is more robust and means I don't have to deal with end-users
reporting this as a problem with the xserver

Popen is not used, but must be built as it is referenced by sdksyms

v2: Update to include fix to Win32TempDir() when TMP is set but TEMP isn't

XXX: Really this should check TMPDIR first, as that's the canonical UNIX way to
set the temporary directory?

v3: Fix implicit-function-declaration of Win32TempDir() in XkbDDXCompileKeymapByNames()

Add a prototype for Win32TempDir()

v4: Improve handling of error conditions with the xkbcomp temporary file

Fix a compilation conditional so that a failure to write to the xkbcomp
temporary file is correctly reported, and reports the problematic filename.

Improve Win32TempDir() so that the directory given by the TEMP and TMP
environment variables is checked for accessibility, so that specifying a
non-existent directory doesn't cause a failure.

Signed-off-by: Jon TURNEY <jon.turney@dronecode.org.uk>