os/osinit.c, os/utils.c: Exclude sigaction code when X_NOT_POSIX
In particular, MinGW doesn't have sigaction, so this patch is needed for building.

No attempt is made to actually install the fatal error signal handler, as MinGW
will terminate the process rather than deliver a fatal signal.

(This is patterned after the use of X_NOT_POSIX around sigaction code in xhost.
X_NOT_POSIX is defined in Xosdefs.h for certain platforms, including WIN32.  It might
be better to just check WIN32)

Signed-off-by: Ryan Pavlik <rpavlik@iastate.edu>
Reviewed-by: Jon TURNEY <jon.turney@dronecode.org.uk>