Do not report the GLX_INTEL_swap_event extension for indirect swrast
Commit 84956ca4 bogusly adds GLX_INTEL_swap_event to the extensions reported
by swrast.  swrast does not support this extension and should not report it,
otherwise toolkits like clutter will only ever SwapBuffers once and wait forever
for an event that's not coming.
(A similar bug for direct swrast is already fixed in mesa commit 25620eb1)

Ensure this kind of bug doesn't occur in future, by rather than using a fixed
list of GLX extensions for swrast, generate the extension list in the same way
as dri and dri2 do.

This still looks a bit wonky: swrast and DRI1 will still always report certain
GLX extensions, as __glXInitExtensionEnableBits() has enabled them, even though
they then subsequently check if the underlying driver supports them (e.g. GLX_SGI_make_current_read)

This has a couple of side-effects on the reported GLX extensions for swrast:
- GLX_SGIS_multisample is now reported on APPLE (I can't work out why this conditional
was added, it dates back to the XFree86 4.4RC3 import)
- GLX_SGIX_visual_select_group is now reported (this is probably pointless but benign)

(Note that this may be papering over the cracks somewhat, as if we do report
GLX_INTEL_swap_event, some clutter apps fail with GLXBadDrawable calling
GLXChangeDrawableAttributes to change the setting of GLX_BUFFER_SWAP_COMPLETE_INTEL_MASK
in the GLX_EVENT_MASK, apparently after the drawable is destroyed, which suggests
a bug with GLXDrawable lifetimes)

Also:
- Don't record GLX_INTEL_swap_event as being required by GLX 1.4, it isn't.
(This data is not currently used)

Signed-off-by: Jon TURNEY <jon.turney@dronecode.org.uk>