Add hook for rootless into miPaintWindow
Revisit b3415187e92960cbff784108b5a3a8d130dc34c5, which shoves some rootless
specific code into miPaintWindow.

Unfortunately as written, this means that a DIX code built with ROOTLESS defined
must link with a DDX which includes ROOTLESS, which makes it impossible to build
rootless supporting (XWin and XQuartz) and non-rootless DDXs at the same time,
which is a pain from a packaging and tinderboxing point of view.

Instead, add a rootless specific hook into miPaintWindow, so DDXs which have the
rootless extension can install the the neccesary checking which rootless needs
to do.  DDXs which don't support rootless won't install the hook, but the DIX
remains compatible with them, so they can be built at the same time.

See [1] for admission that code as it stands is a horrible hack :-)

[1] http://lists.macosforge.org/pipermail/xquartz-dev/2009-September/002551.html