Work-around problems with using Xlib client macros in the server in a 64-bit build.
sizeof(XID) is always 4 in the server, but is 8 in a 64-bit client and coverted
to a 32-bit protocol value (it's 4 in a 32-bit client) (See the _XSERVER64 define)

This affects the size of members of the _XPrivDisplay type used by some macros in
Xlib.h and so means those macros will not give correct answers when we try to use them
in server code.

Switch to using the equivalent functions, which work correctly, as the code in libX11
is not affected by the types used in the server.

A better solution would be to xcb-ize all this code, which also avoids server/client
namespace and header clashes.