Undefine _XSERVER64 in multiwindow WM
The XEvent type suffers from the _XSERVER64 issues noted previously: certain
members are long int in Xlib, but are made int in the server by XSERVER64.

Undefine _XSERVER64 in multiwindow WM, so client code can build correctly on
x86_64.

This is needed to send the WM_KILL ClientMessage using XSendEvent() correctly.

Move winUpdateWindowPosition() to winmultwindowproc.c, as it needs to access the
server Drawable structure correctly, which uses affected types.

(We still play a dangerous game with bringing the WindowPtr type into scope so
we can use it in window shaping code to access the server's shape data.  This
would be better rewritten as client code using ShapeNotify and
XShapeGetRectangles())