Undefine _XSERVER64 in multiwindow WM
Undefine _XSERVER64 in multiwindow WM, so client code can build correctly on
x86_64.

Move winUpdateWindowPosition() to winmultwindowproc.c, as it needs to access the
server Drawable structure correctly, which uses affected types.

This supersedes "Bodge around issues with XSendEvent()" needed to get the
WM_KILL message ClientMessage sent correctly

(We still play a dangerous game with bringing the WindowPtr type into scope so
we can use it in window shaping code to access the server's shape data.  This
would be better rewritten as client code using ShapeNotify and
XShapeGetRectangles())